backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.1
Django==5.2
django-cors-headers==4.7.0
django-stubs==5.2.0
django-stubs-ext==5.2.0
djangorestframework==3.16.0
djangorestframework-stubs==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
easyocr==1.7.2
filelock==3.18.0
fsspec==2025.3.2
idna==3.10
imageio==2.37.0
inflection==0.5.1
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.5
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.2.1
pyclipper==1.3.0.post6
PyJWT==2.9.0
pytesseract==0.3.13
python-bidi==0.6.6
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.25.2
scipy==1.15.2
setuptools==75.8.0
shapely==2.1.0
sqlparse==0.5.3
sympy==1.14.0
tifffile==2025.3.30
torch==2.7.0
torchaudio==2.7.0+cpu
torchvision==0.22.0
types-PyYAML==6.0.12.20250402
types-requests==2.32.0.20250328
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
wheel==0.45.1
